Rename to `ChangeMetricsLink`.
