+++ assets/js/components/KeyMetrics/index.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-export { default as ChangeMetricsCTA } from './ChangeMetricsCTA';
+export { default as ChangeMetricsLink } from './ChangeMetricsLink';
 export { default as KeyMetricsSetupCTAWidget } from './KeyMetricsSetupCTAWidget';
 export { default as MetricTileNumeric } from './MetricTileNumeric';
 export { default as MetricTileTable } from './MetricTileTable';
