+++ tests/backstop/reference/google-site-kit_KeyMetrics_ChangeMetricsLink_0_document_1_medium.png
